% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_runs.R
\name{compare_runs}
\alias{compare_runs}
\title{Compare NONMEM runs estimation results}
\usage{
compare_runs(runs = NULL, estimation_number = NULL)
}
\arguments{
\item{runs}{a list of \code{pmxploit} NONMEM run objects.}

\item{estimation_number}{(optional) integer. A vector of containing the position(s)
of the estimation step(s) to compare in the runs. If \code{estimation_number=-1},
the last estimation step of each run will be selected. Default is \code{NULL} to
consider all estimation steps.}
}
\value{
A data frame.
}
\description{
Returns a table summarizing the results of estimation steps of a list of NONMEM runs.
}
\details{
Results are gathered in a nested data frame, containing one row per estimation step included in the comparison.
Each row contains the following columns: \itemize{
\item \code{RUN_ID}: the run identifying name/number.
\item \code{INFO}: information about the run (problem, number of observations, number of individuals and run duration).
\item \code{FILES}: a tibble containing the control stream name, the dataset name and the run archive path.
\item \code{ESTIMATION}: a nested tibble containing the following columns: \itemize{
\item \code{SUMMARY}: a tibble containing the estimation results summary (method, OFV, termination status, duration...).
\item \code{THETA}: a tibble containing THETA parameters results.
\item \code{OMEGA}: a tibble containing OMEGA parameters results.
\item \code{SIGMA}: a tibble containing SIGMA parameters results.
\item \code{ETABAR}: a tibble containing ETABAR results.
\item \code{SHRINKAGE}: a tibble containing shrinking results.}}
}
\examples{
compare_runs(list(EXAMPLERUN, EXAMPLERUN))
}
