% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_outliers.R
\name{detect_outliers}
\alias{detect_outliers}
\title{Outliers detection}
\usage{
detect_outliers(run, compartment = NULL, residuals = NULL,
  method = "grubbs", grubbs_pvalue_threshold = 0.05,
  boxplot_coefficient = 3, keep_time_zero = FALSE)
}
\arguments{
\item{run}{\code{pmxploit} NONMEM run object.}

\item{compartment}{integer. Number of the compartment of the observations of interest.}

\item{residuals}{character. Column name of the residuals in the output tables.}

\item{method}{character. One of \code{"grubbs"} (for iterative Grubb's test)
or \code{"boxplot"} (for non-parametric detection).}

\item{grubbs_pvalue_threshold}{numeric. p-value threshold for Grubb's test.}

\item{boxplot_coefficient}{numeric. k coefficient for non-parametric test.}

\item{keep_time_zero}{logical. If \code{TRUE}, does not ignore values at initiation (when TIME = 0).
Default is \code{FALSE}.}
}
\value{
A list with the following structure: \itemize{
\item \code{method}: character string of outlier detection method.
\item \code{residuals}: character string of the type of residuals.
\item \code{source}: tibble of the data source.
\item \code{outliers}: tibble of the outliers with 5 columns: \code{ID}, \code{TIME}, \code{CMT}, \code{DV} and
\code{"residuals"}.
}
}
\description{
Detection of observations outliers based on residuals values according to two
different methods: Grubb's test or a non-parametric method.
}
\examples{
EXAMPLERUN \%>\%
 detect_outliers(compartment = 2, residuals = "CWRES", method = "grubbs")

EXAMPLERUN \%>\%
 detect_outliers(compartment = 2, residuals = "CWRES",
                 method = "grubbs", grubbs_pvalue_threshold = 0.10)

EXAMPLERUN \%>\%
 detect_outliers(compartment = 2, residuals = "CWRES", method = "boxplot")
}
