% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EXAMPLERUN}
\alias{EXAMPLERUN}
\title{NONMEM example run}
\format{A NONMEM run object of the class \code{nonmem_run}.}
\usage{
EXAMPLERUN
}
\description{
A NONMEM run example, based on the reference noted below.
}
\references{
Djebli, N., Martinez, JM., Lohan, L. et al. Clin Pharmacokinet (2017) 56: 1155.
\url{https://doi.org/10.1007/s40262-016-0505-1}
}
\keyword{datasets}
