% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{filter.nonmem_run}
\alias{filter.nonmem_run}
\title{Filter a NONMEM run object data}
\usage{
\method{filter}{nonmem_run}(x, ...)
}
\arguments{
\item{x}{An \code{nonmem_run} object.}
}
\description{
Filter a NONMEM run object data
}
\details{
Adds an attribute to the \code{nonmem_run} object so that
every function call on the run (plot, summary, quality criteria) will
first filter the internal data.
}
\examples{
filter(EXAMPLERUN, STUD == 0)

filter(EXAMPLERUN, SEX == 0, STUD == 1)

}
