% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_estimation}
\alias{get_estimation}
\title{Get an estimation step of a run}
\usage{
get_estimation(run, estimation_number = NULL)
}
\arguments{
\item{run}{\code{pmxploit} NONMEM run object.}

\item{estimation_number}{integer. Number of the estimation step.
Default is NULL, returning the last estimation step.}
}
\value{
A list containing the estimation information and results
}
\description{
Get an estimation step of a run from its number
}
