% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{group_by.nonmem_run}
\alias{group_by.nonmem_run}
\title{Group a NONMEM run object data}
\usage{
\method{group_by}{nonmem_run}(x, ...)
}
\arguments{
\item{x}{An \code{nonmem_run} object generated.}
}
\description{
Group a NONMEM run object data
}
\examples{
group_by(EXAMPLERUN, SEX)

group_by(EXAMPLERUN, SEX, STUD)

}
