% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_bootstrap.R
\name{nm_bootstrap}
\alias{nm_bootstrap}
\title{Generates control files and dataset for bootstrap analysis}
\usage{
nm_bootstrap(
  run,
  target_directory = "bootstrap",
  n_sample = 200,
  seed = 1000,
  update_progress = NULL
)
}
\arguments{
\item{run}{\code{pmxploit} NONMEM run object of the source run.}

\item{target_directory}{path of the directory where the generated files will be located.}

\item{n_sample}{integer. Number of runs to generate.}

\item{seed}{integer. Seed number for reproducibility.}

\item{update_progress}{a function useful to monitor the generation process. This function should take a \code{detail} argument.}
}
\value{
A vector containing the paths of the generated control files
}
\description{
Generates control files and dataset for bootstrap analysis
}
\examples{
\dontrun{
nm_bootstrap(pmxploit::EXAMPLERUN)
}
}
