% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_chain.R
\name{nm_chain}
\alias{nm_chain}
\title{Generates control files and dataset for $CHAIN analysis}
\usage{
nm_chain(
  run,
  n_sample = 15,
  iaccept = 0.5,
  ctype = 2,
  degrees_of_freedom = NULL,
  seed = 1000,
  target_directory = "initial",
  update_progress = NULL
)
}
\arguments{
\item{run}{\code{pmxploit} NONMEM run object of the source run.}

\item{n_sample}{integer. Number of runs to generate.}

\item{iaccept}{IACCEPT parameter}

\item{ctype}{CTYPE parameter}

\item{degrees_of_freedom}{DF parameter}

\item{seed}{integer. Seed number for reproducibility.}

\item{target_directory}{path of the directory where the generated files will be located.}

\item{update_progress}{a function useful to monitor the generation process. This function should take a \code{detail} argument.}
}
\value{
A vector containing the paths of the generated control files
}
\description{
Generates control files and dataset for $CHAIN analysis
}
\details{
See NONMEM documentation for details about IACCEPT, CTYPE and DF parameters.
}
\examples{
\notrun{
nm_chain(pmxploit::EXAMPLERUN)
}
}
