% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_check.R
\name{nm_check}
\alias{nm_check}
\title{Performs a NM-TRAN check of the control file}
\usage{
nm_check(control_file, nmcheck_exe = NULL, call_template = NULL)
}
\arguments{
\item{control_file}{Path of the control file}

\item{nmcheck_cmd}{Path of NMTRAN check executable}
}
\value{
A list with control file path and error/error message if any.
}
\description{
Performs a NM-TRAN check of the control file
}
\examples{
\notrun{
nm_check(control_file = "control_file.ctl",
nmcheck_exe = "/path/to/nmtran.exe",
call_template = "cat {control_file} | {nmcheck_exe}")
}
}
