% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_exec.R
\name{nm_exec}
\alias{nm_exec}
\title{Starts a NONMEM run}
\usage{
nm_exec(control_file, run_directory = NULL, nonmem_exe = NULL,
  n_nodes = 1L, parafile = NULL, result_directory = NULL,
  job_id = NULL, archive = TRUE, cleanup = TRUE, quiet = FALSE,
  call_template = NULL, ...)
}
\arguments{
\item{control_file}{path of the control file}

\item{run_directory}{path of the directory where to perform the run}

\item{nonmem_exe}{path of NONMEM executable path}

\item{n_nodes}{number of CPUs}

\item{parafile}{path of a NONMEM parallel processing settings file}

\item{result_directory}{path of the directory where to place the results}

\item{job_id}{(optional) integer run ID}

\item{archive}{logical. If TRUE, compress the result files to a *.tar.gz archive?}

\item{cleanup}{logical. If TRUE, remove NONMEM temporary files}

\item{quiet}{logical. If TRUE, do not print NONMEM output when executing}

\item{call_template}{character string used to construct NONMEM call. Helps creating custom script calls.}
}
\value{
A \code{pmxploit} NONMEM run object.
}
\description{
Starts a NONMEM run
}
\examples{
\\notrun{
nm_exec(control_file = "control_file.ctl", nonmem_exe = "/path/to/nmfe")
}
}
