% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_exec2.R
\name{nm_exec2}
\alias{nm_exec2}
\title{Starts a NONMEM run and do not wait for the results}
\usage{
nm_exec2(control_file, call_template = NULL, ...)
}
\arguments{
\item{control_file}{path of the control file}

\item{call_template}{character string call template used to start the run.}
}
\value{
A character string returned by the system call.
}
\description{
Starts a NONMEM run and do not wait for the results. Useful when NONMEM jobs are run via an HPC scheduler or a cloud solution.
}
\examples{
\dontrun{
nm_exec2(control_file = "control_file.ctl", custom_script = "/path/to/script", call_template = "{custom_script} | {control_file}")
}
}
