% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_jacknife.R
\name{nm_jacknife}
\alias{nm_jacknife}
\title{Generates control files and dataset for jacknife analysis}
\usage{
nm_jacknife(run, target_directory = "jacknife", update_progress = NULL)
}
\arguments{
\item{run}{\code{pmxploit} NONMEM run object of the source run.}

\item{target_directory}{path of the directory where the generated files will be located.}

\item{update_progress}{a function useful to monitor the generation process. This function should take a \code{detail} argument.}
}
\value{
A vector containing the paths of the generated control files
}
\description{
Generates control files and dataset for jacknife analysis
}
\examples{
#' \\notrun{
nm_jacknife(pmxploit::EXAMPLERUN)
}
}
