% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nm_reserved_names}
\alias{nm_reserved_names}
\title{NONMEM reserved names}
\format{A vector of NONMEM reserved names.}
\usage{
nm_reserved_names
}
\description{
NONMEM reserved names related to DATA, PREDPP and NM-TRAN (cf. NONMEM guides).
}
\keyword{datasets}
