% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_simulation.R
\name{nm_simulation}
\alias{nm_simulation}
\title{Generate a NONMEM control stream for simulations from a run}
\usage{
nm_simulation(
  run,
  n_simulations = 100,
  seed = sample.int(1e+09, 1),
  output_table_format = ",1PE11.4",
  output_file = NULL
)
}
\arguments{
\item{run}{\code{pmxploit} NONMEM run object of the source run.}

\item{n_simulations}{integer. Number of simulations (\code{$SIMULATION NSUBPROBLEMS})
to run.}

\item{seed}{integer. Seed number for random number generation.}

\item{output_table_format}{(optional) character. NONMEM format (see NONMEM help) for the simulated
output table. Default is \code{,1PE11.4}, corresponding to a csv format.}

\item{output_file}{(optional) character. File path to write the generated
control stream to. Default is \code{NULL}, not saving any file.}
}
\value{
A character string of the generated control stream code.
}
\description{
Generates a NONMEM control stream for simulations from a source run.
}
\details{
All \code{THETA}, \code{OMEGA} and \code{SIGMA} parameters will be fixed to the source run final estimates.
\code{$ESTIMATION} record(s) will be replaced by a single \code{$SIMULATION} record (with the specified number of simulations).
Simulated output will be saved to a single \code{simtab} file (\code{$TABLE}).
}
\examples{

# Control stream generation from pmxploit::EXAMPLERUN
cs1 <- nm_simulation(EXAMPLERUN, n_simulations = 300)

# see the output
writeLines(cs1)
}
