% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nonmem_txtmsgs}
\alias{nonmem_txtmsgs}
\title{NONMEM text messages}
\format{
A data frame with two variables: \code{id}, \code{message}.
}
\usage{
nonmem_txtmsgs
}
\description{
NONMEM text messages extracted from NONMEM /source/txtmsgs.f90 file.
}
\keyword{datasets}
