% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_nm_control_stream.R
\name{parse_nm_control_stream}
\alias{parse_nm_control_stream}
\title{Parse a NONMEM control stream}
\usage{
parse_nm_control_stream(
  filepath = NULL,
  content = NULL,
  read_initial_values = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{filepath}{character. If \code{content = NULL}, defines the filepath of
the control stream file.}

\item{content}{character. Text content of a control stream file.}

\item{read_initial_values}{logical. If \code{TRUE} (default), parses initial
parameter values from the control stream.}

\item{verbose}{logical. If \code{TRUE}: prints output messages.}
}
\value{
A list representing a NONMEM control stream.

  Object structure: \itemize{
  \item \code{content}: control stream lines.
  \item \code{problem}: problem description.
  \item \code{ignore}: list with: \itemize{
  \item \code{C}: logical. \code{TRUE} if \code{IGNORE=C} \item \code{@}: logical.
   \code{TRUE} if \code{IGNORE=@} \item \code{#}: logical. \code{TRUE} if \code{IGNORE=#}
   \item \code{data}: data frame containing other \code{IGNORE} options.
  } \item \code{ignore_at}: logical, indicating if if \code{IGNORE=@} is specified.
  \item \code{dataset_file}: dataset file name.
  \item \code{subroutine}: NONMEM ADVAN subroutine.
  \item \code{model_compartments}: tibble of the compartments (name and number of each).
  \item \code{subpopulations}: number of subpopulations (for mixture models).
  \item \code{input}: tibble of dataset input columns (name, is dropped?, has a synonymn/alias?).
  \item \code{parameters_definitions}: tibble of the model parameters (\code{THETAs} and \code{ETAs}).
  \item \code{estimations}: successive estimation methods.
  \item \code{tables}: output tables with columns definitions.}
}
\description{
Reads the content of a control stream and extracts informations related to
input data, model, estimation and output.
}
\examples{
\dontrun{
parse_nm_control_stream(filepath = "run_control_stream.con")
}
}
