% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_nm_ext.R
\name{parse_nm_ext}
\alias{parse_nm_ext}
\title{Parse a NONMEM iteration file (*.ext)}
\usage{
parse_nm_ext(filepath = NULL, content = NULL)
}
\arguments{
\item{filepath}{character. If \code{content = NULL}, defines the filepath of
the iteration file.}

\item{content}{character. Text content of an iteration file.}
}
\value{
A tibble with one row per estimation step and 10 columns: \itemize{
\item \code{number}: the number of the estimation step
\item \code{method}: the name of the estimation method
\item \code{goal_function}: the goal function of the estimation method
\item \code{problem}: the goal function of the estimation method
\item \code{subproblem}: subproblem value
\item \code{superproblem1}: superproblem1 value
\item \code{iteration1}: iteration1 value
\item \code{superproblem2}: superproblem2 value
\item \code{iteration2}: iteration2 value
\item \code{iterations}: a tibble with the population parameters and objective function values at each printed out iterations.
}
}
\description{
Reads the content of an iteration file and extracts informations related to the
parameters and the objective function values.
}
\examples{
\dontrun{
parse_nm_ext(filepath = "job_iterations.ext")
}
}
