% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_nm_phi.R
\name{parse_nm_phi}
\alias{parse_nm_phi}
\title{Parse a NONMEM phi file (*.phi)}
\usage{
parse_nm_phi(filepath = NULL, content = NULL)
}
\arguments{
\item{filepath}{character. If \code{content = NULL}, defines the filepath of
the file.}

\item{content}{character. Text content of a phi file.}
}
\value{
A tibble with one row per estimation step and 9 columns: \itemize{
\item \code{number}: the number of the estimation step
\item \code{method}: the name of the estimation method
\item \code{problem}: the goal function of the estimation method
\item \code{subproblem}: subproblem value
\item \code{superproblem1}: superproblem1 value
\item \code{iteration1}: iteration1 value
\item \code{superproblem2}: superproblem2 value
\item \code{iteration2}: iteration2 value
\item \code{individuals}: a tibble with the individual parameter values.
}
}
\description{
Reads the content of a phi file and extracts informations related to the
parameters and the objective function values.
}
\examples{
\dontrun{
parse_nm_phi(filepath = "run.phi")
}
}
