% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_categorical_covariates_distributions.R
\name{plot_categorical_covariates_distributions}
\alias{plot_categorical_covariates_distributions}
\title{Categorical covariates distributions bar charts}
\usage{
plot_categorical_covariates_distributions(
  run,
  covariates = NULL,
  frequency = FALSE,
  order = TRUE,
  drop = FALSE,
  bar_adjustment = "dodge",
  baseline_only = TRUE,
  auto_legend = TRUE
)
}
\arguments{
\item{run}{\code{pmxploit} NONMEM run object.}

\item{covariates}{character vector of categorical covariates names. Default is \code{NULL},
returning all categorical covariates.}

\item{frequency}{logical. Plot frequency instead of count on the y-axis.}

\item{bar_adjustment}{character. \code{ggplot2} \code{position} argument. One
of \code{identity}, \code{dodge}, \code{jitterdodge}, \code{jitter},
\code{nudge} and \code{stack}. (see \code{\link{ggplot2}} documentation).
Default is \code{dodge}.}

\item{baseline_only}{logical. Consider only the baseline (= first) values of the subjects. Default is
\code{TRUE}.}

\item{auto_legend}{logical. When \code{TRUE} (default), default \code{pmxploit} plotting
options are applied (title, labels, axis...).}

\item{split_by}{character. Name of a categorical covariate used to split the
distributions.}
}
\value{
A ggplot2 object.
}
\description{
Plots the distributions of the selected categorical covariates.
}
\examples{
EXAMPLERUN \%>\%
  plot_categorical_covariates_distributions(covariates = c("SEX", "STUD"))

EXAMPLERUN \%>\%
  group_by(SEX) \%>\%
  plot_categorical_covariates_distributions(covariates = "STUD",
                                            frequency = TRUE)
}
