% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_continuous_covariates_distributions.R
\name{plot_continuous_covariates_distributions}
\alias{plot_continuous_covariates_distributions}
\title{Continuous covariates distributions}
\usage{
plot_continuous_covariates_distributions(run, covariates = NULL,
  baseline_only = TRUE, type = "histogram", histogram_bins = 30L,
  boxplot_facets = TRUE, boxplot_drop_unused = FALSE,
  qq_reference_distribution = qnorm, facet_scales = "free",
  overlay_splits = TRUE, auto_legend = TRUE)
}
\arguments{
\item{run}{\code{pmxploit} NONMEM run object.}

\item{covariates}{character vector of continous covariates names. Default is
\code{NULL}, returning all continuous covariates.}

\item{baseline_only}{logical. Consider only the baseline (= first) values of the subjects. Default is
\code{TRUE}.}

\item{type}{character. Type of distribution representation. One of
\code{histogram}, \code{density}, \code{boxplot} and \code{qq} (Q-Q plot of
the Normal distribution). Default is \code{histogram}.}

\item{histogram_bins}{integer. If \code{type = "histogram"}, sets the number
of bins. Default is 30.}

\item{boxplot_facets}{logical. If \code{type = "boxplot"}, represent
boxplots in different facets when \code{TRUE}.}

\item{boxplot_drop_unused}{logical. If \code{type = "boxplot"}, drops unused
factor levels when \code{TRUE}.}

\item{facet_scales}{character. \code{ggplot2} facet scales. Default is
\code{"free"}.}

\item{overlay_splits}{logical. If splitting is applied, \code{TRUE} will
overlay splits on a same graph while \code{FALSE} will plot them on
different facets. Default is \code{TRUE}.}

\item{auto_legend}{logical. When \code{TRUE} (default), default \code{pmxploit} plotting
options are applied (title, labels, axis...).}

\item{split_by}{character vector. Names of categorical covariates used to
split the distributions.}
}
\value{
A ggplot2 object.
}
\description{
Plots the distributions of the selected continuous covariates.
}
\examples{
EXAMPLERUN \%>\% plot_continuous_covariates_distributions()
EXAMPLERUN \%>\% plot_continuous_covariates_distributions(type = "density")
EXAMPLERUN \%>\% plot_continuous_covariates_distributions(type = "boxplot")
EXAMPLERUN \%>\% plot_continuous_covariates_distributions(type = "qq")

EXAMPLERUN \%>\%
  group_by(SEX) \%>\%
  plot_continuous_covariates_distributions(type = "boxplot", covariates = c("AGE", "WT"))

EXAMPLERUN \%>\%
  group_by(SEX, STATIN) \%>\%
  plot_continuous_covariates_distributions(type = "boxplot", covariates = c("AGE", "WT"))

EXAMPLERUN \%>\%
  group_by(SEX) \%>\%
  plot_continuous_covariates_distributions(type = "histogram", covariates = c("AGE", "WT"))

EXAMPLERUN \%>\%
  group_by(SEX) \%>\%
  plot_continuous_covariates_distributions(type = "histogram",
                                           covariates = c("AGE", "WT"),
                                           overlay_splits = FALSE)

EXAMPLERUN \%>\%
  group_by(SEX) \%>\%
  plot_continuous_covariates_distributions(type = "density")
}
