% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_convergence.R
\name{plot_convergence}
\alias{plot_convergence}
\title{Convergence plot}
\usage{
plot_convergence(
  run,
  estimation_number = NULL,
  parameters = NULL,
  auto_legend = TRUE
)
}
\arguments{
\item{run}{\code{pmxploit} NONMEM run object.}

\item{estimation_number}{integer. Number of the estimation step. Default is
\code{NULL}, plotting all the steps.}

\item{parameters}{character. Either one of \code{"theta"}, \code{"omega"} or
\code{"sigma"} or a character vector of population parameters names. Default
is \code{NULL} plotting all population parameters.}

\item{auto_legend}{logical. When \code{TRUE} (default), default \code{pmxploit} plotting
options are applied (title, labels, axis...).}
}
\value{
A ggplot2 object.
}
\description{
Plots the value, at each printed out iteration, of the population parameters (\code{THETA}, \code{OMEGA} or
\code{SIGMA}) during an estimation step. Fixed parameters are not plotted.
}
\examples{
EXAMPLERUN \%>\% plot_convergence()

EXAMPLERUN \%>\% plot_convergence(parameters = "theta")
EXAMPLERUN \%>\% plot_convergence(parameters = "omega")

EXAMPLERUN \%>\% plot_convergence(parameters = c(CL_pop = "THETA3"))
}
