% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dv_vs_predictions.R
\name{plot_dv_vs_predictions}
\alias{plot_dv_vs_predictions}
\title{Dependent variables vs Predictions plot}
\usage{
plot_dv_vs_predictions(run = NULL, compartment = NULL, dv = "DV",
  predictions = "PRED", log_dv = FALSE, facetted = TRUE,
  smoothing_method = NULL, smoothing_se = TRUE, identity_line = TRUE,
  x_scale = "linear", y_scale = "linear", logticks_annotation = TRUE,
  facet_scales = "free", transparency = FALSE,
  keep_time_zero = FALSE, auto_legend = TRUE)
}
\arguments{
\item{run}{\code{pmxploit} NONMEM run object.}

\item{compartment}{integer. Number of the compartment of the dependent variable.}

\item{dv}{character. Name of the dependent variable column in the output tables.
Default is \code{"DV"}.}

\item{predictions}{character vector. Name of the predictions column(s) in the
dataset. Default is \code{"PRED"} for both population predictions and
individual predictions.#'}

\item{log_dv}{logical. Set it to \code{TRUE} if the source dependent variable are in log_scale.}

\item{facetted}{logical. If \code{TRUE} (default), plots each "DV vs predictions" subset inside a separate facet.}

\item{smoothing_method}{character. \code{ggplot2} smoothing method.}

\item{smoothing_se}{logical. Display confidence interval around smooth?}

\item{identity_line}{logical. Draws an identity line. Default is \code{TRUE}.}

\item{x_scale}{character. X-axis scale, one of \code{"linear"} or
\code{"log"}. Default is \code{"linear"}.}

\item{y_scale}{character. Y-axis scale, one of \code{"linear"} or
\code{"log"}. Default is \code{"linear"}.}

\item{logticks_annotation}{logical. If \code{x_scale} and/or \code{y_scale} are \code{"log"}, adds ticks annotation on the axis.}

\item{facet_scales}{character. \code{ggplot2} facet scales. Default is
\code{"free"}.}

\item{transparency}{logical. Plot scatterplot dots with transparency, useful
to avoid overplotting with large datasets. Default is {FALSE}.#'}

\item{keep_time_zero}{logical. If \code{TRUE}, does not ignore values at initiation (when TIME = 0).
Default is \code{FALSE}.}

\item{auto_legend}{logical. When \code{TRUE} (default), default \code{pmxploit} plotting
options are applied (title, labels, axis...).}

\item{split_by}{character vector. Name of categorical covariates used to split the
data.}
}
\value{
A ggplot2 object.
}
\description{
Dependent variables vs Predictions plot
}
\examples{

EXAMPLERUN \%>\%
  plot_dv_vs_predictions(compartment = 2, predictions = "PRED")

EXAMPLERUN \%>\%
  group_by(CMT) \%>\%
  plot_dv_vs_predictions(compartment = 2:3, predictions = "PRED")

EXAMPLERUN \%>\%
  plot_dv_vs_predictions(compartment = 2, predictions = "PRED", x_scale = "log", y_scale = "log")

EXAMPLERUN \%>\%
  plot_dv_vs_predictions(compartment = 2, predictions = "PRED", smoothing_method = "lm")
EXAMPLERUN \%>\%
  plot_dv_vs_predictions(compartment = 2, predictions = "PRED", smoothing_method = "loess")
EXAMPLERUN \%>\%
  plot_dv_vs_predictions(compartment = 2, predictions = "PRED", identity_line = FALSE)
EXAMPLERUN \%>\%
  plot_dv_vs_predictions(compartment = 2, predictions = "PRED", transparency = TRUE)

EXAMPLERUN \%>\%
  group_by(STUD) \%>\%
  plot_dv_vs_predictions(compartment = 2, predictions = "PRED")
}
