% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_individual_profiles.R
\name{plot_individual_profiles}
\alias{plot_individual_profiles}
\title{Individual profiles (observations, predictions) versus independent variable}
\usage{
plot_individual_profiles(
  run,
  ids = NULL,
  compartment,
  predictions = "PRED",
  idv = "TIME",
  log_dv = FALSE,
  predictions_dots = TRUE,
  show_observations = TRUE,
  categorical_covariate = NULL,
  x_scale = "linear",
  y_scale = "linear",
  logticks_annotation = TRUE,
  facetted = TRUE,
  facet_scales = "free",
  n_row = NULL,
  n_col = NULL,
  keep_time_zero = FALSE,
  auto_legend = TRUE
)
}
\arguments{
\item{run}{\code{pmxploit} NONMEM run object.}

\item{ids}{integer vector of the IDs of the individuals to plot. Default is
\code{NULL}, plotting all the individuals.}

\item{compartment}{integer. Number of the compartment of the dependent variable.}

\item{predictions}{character vector. Name of the predictions column(s) in the
dataset. Default is \code{"PRED"} for both population predictions and
individual predictions.}

\item{idv}{character. Name of the independent variable column to plot on the x-axis. Default is \code{"TIME"}.}

\item{log_dv}{logical. Set it to \code{TRUE} if the source dependent variable are in log_scale.}

\item{predictions_dots}{logical. If \code{TRUE} plot predictions as dots on top of the profile lines.}

\item{show_observations}{logical. If \code{TRUE} plot observations.}

\item{categorical_covariate}{character. A categorical covariate to split the data. May be useful for covariates that vary within an individual.}

\item{x_scale}{character. X-axis scale, one of \code{"linear"} or
\code{"log"}. Default is \code{"linear"}.}

\item{y_scale}{character. Y-axis scale, one of \code{"linear"} or
\code{"log"}. Default is \code{"linear"}.}

\item{logticks_annotation}{logical. If \code{x_scale} and/or \code{y_scale} are \code{"log"}, adds ticks annotation on the axis.}

\item{facetted}{logical. If \code{TRUE} (default), plots each individual inside a separate facet.}

\item{facet_scales}{character. \code{ggplot2} facet scales. Default is
\code{"free"}.}

\item{n_row}{integer. Number of rows of facets.}

\item{n_col}{integer. Number of columns of facets.}

\item{keep_time_zero}{logical. If \code{TRUE}, does not ignore values at initiation (when TIME = 0).
Default is \code{FALSE}.}

\item{auto_legend}{logical. When \code{TRUE} (default), default \code{pmxploit} plotting
options are applied (title, labels, axis...).}
}
\value{
A ggplot2 object.
}
\description{
Individual profiles (observations, predictions) versus independent variable
}
\examples{
EXAMPLERUN \%>\% plot_individual_profiles(ids = 1:4, compartment = 2, predictions = "PRED")

EXAMPLERUN \%>\% plot_individual_profiles(compartment = 2, predictions = "PRED", facetted = FALSE)
}
