% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_observed_profiles.R
\name{plot_observed_profiles}
\alias{plot_observed_profiles}
\title{Spaghetti plot}
\usage{
plot_observed_profiles(run, compartment = NULL, idv = "TIME",
  ids = NULL, show_mdv = TRUE, log_dv = FALSE,
  mean_profiles = FALSE, x_scale = "linear", y_scale = "linear",
  logticks_annotation = TRUE, facetted = TRUE, facet_scales = "free",
  transparency = FALSE, auto_legend = TRUE)
}
\arguments{
\item{run}{\code{pmxploit} NONMEM run object.}

\item{compartment}{integer. Number of the compartment of the dependent variable.}

\item{idv}{character. Name of the column used as independent variable. Default is
\code{"TIME"}.}

\item{ids}{integer vector of the IDs of the individuals to plot. Default is
\code{NULL}, plotting all the individuals.}

\item{log_dv}{logical. Set it to \code{TRUE} is dependent variable is in
log-scale in the dataset. Default is \code{FALSE}.}

\item{mean_profiles}{logical. If \code{TRUE}, summarizes individual profiles
by the mean profile (in each group if \code{split_by} is specified).}

\item{x_scale}{character. X-axis scale, one of \code{"linear"} or
\code{"log"}. Default is \code{"linear"}.}

\item{y_scale}{character. Y-axis scale, one of \code{"linear"} or
\code{"log"}. Default is \code{"linear"}.}

\item{logticks_annotation}{logical. If \code{x_scale} and/or \code{y_scale} are \code{"log"}, adds ticks annotation on the axis.}

\item{facetted}{logical. If \code{TRUE} (default) and \code{split_by}
is specified, plot each group in a separate facet. Otherwise, draw groups
in different colours.}

\item{facet_scales}{character. \code{ggplot2} facet scales. Default is
\code{"free"}.}

\item{transparency}{logical. Plot scatterplot dots with transparency, useful
to avoid overplotting with large datasets. Default is {FALSE}.#'}

\item{auto_legend}{logical. When \code{TRUE} (default), default \code{pmxploit} plotting
options are applied (title, labels, axis...).}
}
\value{
A ggplot2 object.
}
\description{
Plot of observed profiles versus an independent variable (e.g. TIME or TAD).
}
\examples{
EXAMPLERUN \%>\% plot_observed_profiles(compartment = 2)

EXAMPLERUN \%>\% group_by(CMT) \%>\% plot_observed_profiles(compartment = 2:3)

EXAMPLERUN \%>\% group_by(SEX) \%>\% plot_observed_profiles(compartment = 2)
EXAMPLERUN \%>\% group_by(SEX) \%>\% plot_observed_profiles(compartment = 2, facetted = FALSE)

EXAMPLERUN \%>\% group_by(SEX) \%>\% plot_observed_profiles(compartment = 2, mean_profiles = TRUE)
EXAMPLERUN \%>\% group_by(SEX) \%>\%
  plot_observed_profiles(compartment = 2, mean_profiles = TRUE, facetted = FALSE)

EXAMPLERUN \%>\%
  group_by(SEX) \%>\%
  plot_observed_profiles(compartment = 2, y_scale = "log", facetted = FALSE)
}
