% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ofv.R
\name{plot_ofv}
\alias{plot_ofv}
\title{Objective function plot}
\usage{
plot_ofv(run, estimation_number = NULL, auto_legend = TRUE)
}
\arguments{
\item{run}{\code{pmxploit} NONMEM run object.}

\item{estimation_number}{integer. Number of the estimation step. Default is
\code{NULL}, plotting the last estimation.}

\item{auto_legend}{logical. When \code{TRUE} (default), default \code{pmxploit} plotting
options are applied (title, labels, axis...).}
}
\value{
A ggplot2 object.
}
\description{
Plots the value, at each printed out iteration, of objective function during an estimation step.
}
\examples{
EXAMPLERUN \%>\% plot_ofv()

EXAMPLERUN \%>\% plot_ofv(estimation_number = 1)
}
