% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_parameters_distributions.R
\name{plot_parameters_distributions}
\alias{plot_parameters_distributions}
\title{Individual parameters distributions}
\usage{
plot_parameters_distributions(
  run,
  parameters = NULL,
  baseline_only = TRUE,
  type = "histogram",
  histogram_bins = 30L,
  histogram_empirical_density = FALSE,
  histogram_reference_distribution = list(fun = dnorm, args = list(mean = 0, sd = 1)),
  qq_reference_distribution = list(fun = qnorm, args = list(mean = 0, sd = 1)),
  boxplot_facets = TRUE,
  boxplot_drop_unused = FALSE,
  facet_scales = "free",
  overlay_splits = TRUE,
  auto_legend = TRUE
)
}
\arguments{
\item{run}{\code{pmxploit} NONMEM run object.}

\item{parameters}{character vector of parameters names. Default is \code{NULL},
returning all individual parameters (random and post-hoc).}

\item{baseline_only}{logical. Consider only the baseline (= first) values of the subjects. Default is
\code{TRUE}.}

\item{type}{character. Type of distribution representation. One of
\code{histogram}, \code{density}, \code{boxplot} and \code{qq} (Q-Q plot of
the Normal distribution). Default is \code{histogram}.}

\item{histogram_bins}{integer. If \code{type = "histogram"}, sets the number
of bins. Default is 30.}

\item{boxplot_facets}{logical. If \code{type = "boxplot"}, represent
boxplots in different facets when \code{TRUE}.}

\item{boxplot_drop_unused}{logical. If \code{type = "boxplot"}, drops unused
factor levels when \code{TRUE}.}

\item{facet_scales}{character. \code{ggplot2} facet scales. Default is
\code{"free"}.}

\item{overlay_splits}{logical. If splitting is applied, \code{TRUE} will
overlay splits on a same graph while \code{FALSE} will plot them on
different facets. Default is \code{TRUE}.}

\item{auto_legend}{logical. When \code{TRUE} (default), default \code{pmxploit} plotting
options are applied (title, labels, axis...).}
}
\value{
A ggplot2 object.
}
\description{
Plots the distributions of the selected individual parameters.
}
\examples{
EXAMPLERUN \%>\% plot_parameters_distributions()
EXAMPLERUN \%>\% plot_parameters_distributions(type = "density")
EXAMPLERUN \%>\% plot_parameters_distributions(type = "boxplot")
EXAMPLERUN \%>\% plot_parameters_distributions(type = "qq")

EXAMPLERUN \%>\% plot_parameters_distributions(parameters = "eta")

EXAMPLERUN \%>\%
  plot_parameters_distributions(parameters = "ETCL",
                                histogram_empirical_density = TRUE,
                                histogram_reference_distribution = list(fun = dnorm,
                                                                        args = list(mean = 0, sd = 1)))
}
