% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_parameters_vs_categorical_covariates.R
\name{plot_parameters_vs_categorical_covariates}
\alias{plot_parameters_vs_categorical_covariates}
\title{Plot parameters vs categorical covariates correlations}
\usage{
plot_parameters_vs_categorical_covariates(run, parameters = NULL,
  covariates = NULL, baseline_only = TRUE, auto_legend = TRUE)
}
\arguments{
\item{run}{\code{pmxploit} NONMEM run object.}

\item{parameters}{character vector of parameters names. Default is \code{NULL},
returning all individual parameters (random and post-hoc).}

\item{covariates}{character vector of categorical covariates names. Default is \code{NULL},
returning all categorical covariates.}

\item{baseline_only}{logical. Consider only the baseline (= first) values of the subjects. Default is
\code{TRUE}.}

\item{auto_legend}{logical. When \code{TRUE} (default), default \code{pmxploit} plotting
options are applied (title, labels, axis...).}
}
\value{
A ggplot2 object.
}
\description{
Plots the selected parameters and categorical covariates correlations.
}
\examples{
EXAMPLERUN \%>\%
  plot_parameters_vs_categorical_covariates(parameters = c("ETCL", "ETV1", "ETKA"),
                                            covariates = c("ISC", "SEX", "STUD"))
}
