% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_parameters_vs_continuous_covariates.R
\name{plot_parameters_vs_continuous_covariates}
\alias{plot_parameters_vs_continuous_covariates}
\title{Parameters vs continuous covariates correlations}
\usage{
plot_parameters_vs_continuous_covariates(run, parameters = NULL,
  covariates = NULL, type = "heatmap", correlation_method = NULL,
  auto_order = TRUE, baseline_only = TRUE, smoothing_method = NULL,
  smoothing_se = TRUE, fixed_ratio = TRUE, auto_legend = TRUE)
}
\arguments{
\item{run}{\code{pmxploit} NONMEM run object.}

\item{parameters}{character vector of parameters names. Default is \code{NULL},
returning all individual parameters (random and post-hoc).}

\item{covariates}{character vector of continous covariates names. Default is
\code{NULL}, returning all continuous covariates.}

\item{type}{character. Type of distribution representation. One of
\code{histogram}, \code{density}, \code{boxplot} and \code{qq} (Q-Q plot of
the Normal distribution). Default is \code{histogram}.}

\item{correlation_method}{a character string indicating which correlation
coefficient (or covariance) is to be computed. One of \code{"pearson"}
(default), \code{"kendall"}, or \code{"spearman"}: can be abbreviated. If
\code{type = "scatterplot"}, \code{"pearson"} method will be used.}

\item{auto_order}{logical. If \code{type = "heatmap"}, auto-orders the
matrix with respect to the distances between values. Default is
\code{TRUE}.}

\item{baseline_only}{logical. Consider only the baseline (= first) values of the subjects. Default is
\code{TRUE}.}

\item{smoothing_method}{character. If \code{type = "scatterplot"}, corresponds
to a \code{ggplot2} smoothing method.}

\item{smoothing_se}{logical. Display confidence interval around smooth?}

\item{fixed_ratio}{logical. If \code{type = "heatmap"}, plot scaled to a
1:1 ratio. Default is \code{TRUE}.}

\item{auto_legend}{logical. When \code{TRUE} (default), default \code{pmxploit} plotting
options are applied (title, labels, axis...).}
}
\value{
A ggplot2 object.
}
\description{
Returns a plot of the selected parameters and continuous covariates correlations.
}
\examples{

params <- c("ETCL", "ETV1", "KINT", "KDEG")
covs <- c("AGE", "WT", "BSLDLC", "FBSPCSK")

EXAMPLERUN \%>\%
  plot_parameters_vs_continuous_covariates(parameters = params, covariates = covs, type = "heatmap")+
  ggplot2::theme(legend.position = "bottom")

EXAMPLERUN \%>\%
  plot_parameters_vs_continuous_covariates(parameters = params, covariates = covs, type = "scatterplot", smoothing_method = "lm")
EXAMPLERUN \%>\%
  plot_parameters_vs_continuous_covariates(parameters = params, covariates = covs, type = "scatterplot", smoothing_method = "loess")
EXAMPLERUN \%>\%
  group_by(SEX) \%>\%
  plot_parameters_vs_continuous_covariates(parameters = params, covariates = covs, type = "scatterplot")
}
