% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_residuals.R
\name{plot_residuals}
\alias{plot_residuals}
\title{Residuals plot}
\usage{
plot_residuals(run, compartment = NULL, idv = "TIME", residuals,
  absolute_residuals = FALSE, keep_time_zero = FALSE,
  type = "scatterplot", smoothing_method = NULL, smoothing_se = TRUE,
  reference_value = NULL, histogram_bins = 30L,
  histogram_empirical_density = TRUE,
  histogram_reference_distribution = list(fun = dnorm, args = list(mean =
  0, sd = 1)), qq_reference_distribution = list(fun = qnorm, args =
  list(mean = 0, sd = 1)), x_scale = "linear", y_scale = "linear",
  facet_scales = "free", transparency = FALSE, auto_legend = TRUE)
}
\arguments{
\item{run}{\code{pmxploit} NONMEM run object.}

\item{compartment}{integer vector of the numbers of the compartments of residuals.}

\item{idv}{character. Name of the column used as independent variable. Default is
\code{"TIME"}.}

\item{residuals}{character vector of column names of residuals.}

\item{absolute_residuals}{logical. If \code{TRUE} residuals are transformed
to absolute values.}

\item{keep_time_zero}{logical. If \code{TRUE}, does not ignore values at initiation (when TIME = 0).
Default is \code{FALSE}.}

\item{type}{character. Type of representation of the residuals. One of
\code{scatterplot}, \code{histogram} and \code{qq} (Q-Q plot of a distribution of reference).
Default is \code{scatterplot}.}

\item{smoothing_method}{character. \code{ggplot2} smoothing method.}

\item{smoothing_se}{logical. Display confidence interval around smooth?}

\item{reference_value}{numeric. Draws reference lines at +/- a particular
value.}

\item{histogram_bins}{integer. If \code{type = "histogram"}, sets the number
of bins. Default is 30.}

\item{histogram_empirical_density}{logical. If \code{TRUE}, draws the
empirical density over the histogram.}

\item{histogram_reference_distribution}{logical. If \code{TRUE}, draws a
theoretical density over the histogram.}

\item{x_scale}{character. X-axis scale, one of \code{"linear"} or
\code{"log"}. Default is \code{"linear"}.}

\item{y_scale}{character. Y-axis scale, one of \code{"linear"} or
\code{"log"}. Default is \code{"linear"}.}

\item{facet_scales}{character. \code{ggplot2} facet scales. Default is
\code{"free"}.}

\item{transparency}{logical. Plot scatterplot dots with transparency, useful
to avoid overplotting with large datasets. Default is {FALSE}.#'}

\item{auto_legend}{logical. When \code{TRUE} (default), default \code{pmxploit} plotting
options are applied (title, labels, axis...).}
}
\value{
A ggplot2 object.
}
\description{
Plot residuals (e.g. \code{CWRES}, \code{IWRES}, \code{NPDE}) as either a
scatterplot (versus an independent variable), an histogram (distribution) or a quantile-quantile plot.
}
\examples{
EXAMPLERUN \%>\%
  plot_residuals(compartment = 2, idv = c("TIME", "PRED"), residuals = "CWRES")

EXAMPLERUN \%>\%
  group_by(CMT) \%>\%
  plot_residuals(compartment = 2:3, idv = c("TIME", "PRED"), residuals = "CWRES")

EXAMPLERUN \%>\%
  plot_residuals(compartment = 2, idv = c("TIME", "PRED"),
                 residuals = "CWRES", smoothing_method = "lm")

EXAMPLERUN \%>\%
  plot_residuals(compartment = 2, idv = c("TIME", "PRED"),
                 residuals = "CWRES", type = "histogram")
EXAMPLERUN \%>\%
  plot_residuals(compartment = 2, idv = c("TIME", "PRED"),
                 residuals = "CWRES", type = "qq")

EXAMPLERUN \%>\%
  group_by(STUD) \%>\%
  plot_residuals(compartment = 2, idv = c("TIME", "PRED"),
                 residuals = "CWRES", type = "qq")
}
