% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{print.nonmem_control_stream}
\alias{print.nonmem_control_stream}
\title{Print a NONMEM control_stream object}
\usage{
\method{print}{nonmem_control_stream}(x, ...)
}
\arguments{
\item{x}{A \code{nonmem_control_stream} object generated.}
}
\description{
Print a NONMEM control_stream object
}
\examples{
# Using the print function
print(EXAMPLERUN$control_stream)

# Or
EXAMPLERUN$control_stream

}
