% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{print.nonmem_run}
\alias{print.nonmem_run}
\title{Print a NONMEM run object}
\usage{
\method{print}{nonmem_run}(x, ...)
}
\arguments{
\item{x}{A \code{nonmem_run} object generated.}
}
\description{
Print a NONMEM run object
}
\examples{
# Using the print function
print(EXAMPLERUN)

# Or
EXAMPLERUN

}
