% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_criteria.R
\name{quality_criteria}
\alias{quality_criteria}
\title{Quality criteria (QC) computations}
\usage{
quality_criteria(
  run,
  predictions = "PRED",
  log_data = FALSE,
  alpha = 0.05,
  drop_empty_splits = FALSE
)
}
\arguments{
\item{run}{\code{pmxploit} NONMEM run object.}

\item{predictions}{character. Name of the predictions column in the result
tables. Default is \code{"PRED"} for population predictions.}

\item{alpha}{numeric. Alpha risk. Used for bias confidence interval
computation and T-test comparing observations and predictions. Default is
\code{5\%}.}

\item{drop_empty_splits}{logical. Drop empty split groups from QC results.}
}
\value{
A list containing the quality criteria analysis results.
}
\description{
Computation of common quality criteria for evaluation of model predictions
performance: standard qc (average fold error, maximal error), bias (mean predictions error), precision
(root mean square error), Student's t-Test, correlation test and linear
regression.
}
\details{
For quality criteria computations, residuals are computed based on
  the formula:

  \eqn{pred_err_i = pred_i - obs_i} \itemize{ \item Standard QC \itemize{ \item Maximal
  Error:
  \deqn{ME=max(|obs-pred|)}
  \item Absolute Average Fold Error:

  \deqn{AAFE=10^(mean(log10(pred/obs)))}
  For reference, see \url{https://www.ncbi.nlm.nih.gov/pubmed/26696327}}
  \item Bias: Mean Prediction Error (MPE) \itemize{ \item Absolute:
  \code{mean(pred_err)} \item Confidence interval for a given \code{alpha} \item
  Relative: \code{mean(pred_err/obs)} } \item Precision: Root Mean Square
  Error (RMSE) \itemize{ \item Absolute: Student's t-Test estimate of
  \code{t.test((obs - pred)^2)} \item Confidence interval for a given
  \code{alpha} \item Relative: \code{rmse/mean(obs)} } \item Student's
  t-Test: observations vs predictions (paired, two-sided)

  Returns estimate, statistic, p-value, degrees of freedom (parameter) and
  confidence interval given \code{alpha}. \item Correlation test between
  observations and predictions

  Returns estimate, statistic, p-value, degrees of freedom (parameter) and
  confidence interval given \code{alpha}. \item Linear regression: \eqn{pred
  = intercept + slope * obs}

  Returns intercept and slope estimates, standard errors, statistics and
  p-values given \code{alpha}. }
}
\examples{
EXAMPLERUN \%>\% quality_criteria()

EXAMPLERUN \%>\% quality_criteria(alpha = 0.01)

EXAMPLERUN \%>\% group_by(SEX) \%>\% quality_criteria()

EXAMPLERUN \%>\% filter(STUD == 2) \%>\% quality_criteria()

EXAMPLERUN \%>\% filter(STUD != 2) \%>\% quality_criteria()

EXAMPLERUN \%>\% filter(AGE <= 24) \%>\% quality_criteria()
}
