% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{recode_categorical_covariates}
\alias{recode_categorical_covariates}
\title{Recode categorical covariates levels}
\usage{
recode_categorical_covariates(run, covariates_levels = NULL)
}
\arguments{
\item{run}{\code{pmxploit} NONMEM run object.}

\item{covariates_levels}{list of categorical covariates, each list item
contains the corresponding categorical levels.}
}
\value{
A NONMEM run object.
}
\description{
Change the name of categorical covariates levels.
}
\examples{
recoded_EXAMPLERUN <-
 recode_categorical_covariates(run = EXAMPLERUN,
                               covariates_levels =
                                    list(STUD = list(`Study 1` = "1",
                                                      `Study 2` = "2",
                                                      `Study 3` = "3")))

# before
EXAMPLERUN \%>\%
  group_by(STUD) \%>\%
  plot_parameters_distributions(type = "boxplot", parameters = "CL")
# after
recoded_EXAMPLERUN \%>\%
  group_by(STUD) \%>\%
  plot_parameters_distributions(type = "boxplot", parameters = "CL")
}
