% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rename_covariates}
\alias{rename_covariates}
\title{Rename model covariates}
\usage{
rename_covariates(run, covariates = NULL)
}
\arguments{
\item{run}{\code{pmxploit} NONMEM run object.}

\item{covariates}{named character vector with new covariates names as items names and
old covariates names as values.}
}
\value{
A NONMEM run object.
}
\description{
Rename model covariates
}
\examples{
renamed_EXAMPLERUN <- EXAMPLERUN \%>\%
rename_covariates(covariates = c("Weight" = "WT", "Study" = "STUD"))

EXAMPLERUN \%>\%
  group_by(STUD) \%>\%
  plot_continuous_covariates_distributions(type = "boxplot")
renamed_EXAMPLERUN \%>\%
  group_by(STUD) \%>\%
  plot_continuous_covariates_distributions(type = "boxplot")

# splitting also works using the original column name
renamed_EXAMPLERUN \%>\%
  group_by(STUD) \%>\%
  plot_continuous_covariates_distributions(type = "boxplot")
}
