% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_compartments_units}
\alias{set_compartments_units}
\title{Set compartment units}
\usage{
set_compartments_units(run, units = NULL)
}
\arguments{
\item{run}{\code{pmxploit} NONMEM run object.}

\item{units}{}
}
\value{
A NONMEM run object.
}
\description{
Set compartment units
}
\examples{

EXAMPLERUN \%>\%
 set_compartments_units(units = c("2" = "nM", "3" = "nM"))
}
