% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{slice.nonmem_run}
\alias{slice.nonmem_run}
\title{Slice a NONMEM run object data}
\usage{
\method{slice}{nonmem_run}(x, ...)
}
\arguments{
\item{x}{An \code{nonmem_run} object generated.}
}
\description{
Slice a NONMEM run object data
}
\examples{
slice(EXAMPLERUN, 1:10)

}
