% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_categorical_covariates.R
\name{summarize_categorical_covariates}
\alias{summarize_categorical_covariates}
\title{Categorical covariates summary}
\usage{
summarize_categorical_covariates(run, covariates = NULL, baseline_only = TRUE)
}
\arguments{
\item{run}{\code{pmxploit} NONMEM run object.}

\item{covariates}{character vector of categorical covariates names. Default is \code{NULL},
returning all categorical covariates.}

\item{baseline_only}{logical. Consider only the baseline (= first) values of the subjects. Default is
\code{TRUE}.}
}
\value{
A data frame.
}
\description{
Summarizes the selected categorical covariates, returning counts and frequencies
of each levels.
}
\examples{
EXAMPLERUN \%>\% summarize_categorical_covariates()
}
