% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_continuous_covariates.R
\name{summarize_continuous_covariates}
\alias{summarize_continuous_covariates}
\title{Continuous covariates distributions summary}
\usage{
summarize_continuous_covariates(run, covariates = NULL,
  quantiles = c(0.05, 0.25, 0.75, 0.95), baseline_only = TRUE)
}
\arguments{
\item{run}{\code{pmxploit} NONMEM run object.}

\item{covariates}{character vector of continous covariates names. Default is
\code{NULL}, returning all continuous covariates.}

\item{quantiles}{numeric vector of quantiles. Default are 5th, 25th, 75th and 95th percentiles.}

\item{baseline_only}{logical. Consider only the baseline (= first) values of the subjects. Default is
\code{TRUE}.}
}
\value{
A data frame.
}
\description{
Summarizes the selected continuous covariates: number of values, mean, median,
quantiles and range.
}
\examples{
EXAMPLERUN \%>\% summarize_continuous_covariates()

EXAMPLERUN \%>\% summarize_continuous_covariates(quantiles = seq(0.05, 0.95, 0.05))
EXAMPLERUN \%>\% summarize_continuous_covariates(quantiles = NULL)

EXAMPLERUN \%>\% group_by(STUD) \%>\%  summarize_continuous_covariates()
EXAMPLERUN \%>\% group_by(STUD) \%>\% summarize_continuous_covariates(quantiles = NULL)
}
