% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_covariates_correlations.R
\name{summarize_covariates_correlations}
\alias{summarize_covariates_correlations}
\title{Continuous covariates correlations summary}
\usage{
summarize_covariates_correlations(run, covariates = NULL,
  baseline_only = TRUE, correlation_method = NULL, auto_order = TRUE)
}
\arguments{
\item{run}{\code{pmxploit} NONMEM run object.}

\item{covariates}{character vector of continous covariates names. Default is
\code{NULL}, returning all continuous covariates.}

\item{baseline_only}{logical. Consider only the baseline (= first) values of the subjects. Default is
\code{TRUE}.}

\item{correlation_method}{a character string indicating which correlation
coefficient (or covariance) is to be computed. One of \code{"pearson"}
(default), \code{"kendall"}, or \code{"spearman"}: can be abbreviated. If
\code{type = "scatterplot"}, \code{"pearson"} method will be used.}

\item{auto_order}{logical. If \code{type = "heatmap"}, auto-orders the
matrix with respect to the distances between values. Default is
\code{TRUE}.}

\item{split_by}{character. Categorical to split the summary by.}
}
\value{
A matrix.
}
\description{
Returns a matrix of the selected continuous covariates correlations.
}
\examples{
EXAMPLERUN \%>\% summarize_covariates_correlations()
EXAMPLERUN \%>\% summarize_covariates_correlations(auto_order = FALSE)
}
