% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_population.R
\name{summarize_etabars}
\alias{summarize_etabars}
\title{Summarize ETA bars}
\usage{
summarize_etabars(run, estimation_number = NULL)
}
\arguments{
\item{run}{\code{pmxploit} NONMEM run object.}

\item{estimation_number}{integer. Number of the estimation step.
Default is NULL, returning the last estimation step.}
}
\value{
A data frame with one row per parameter and the following columns:
\itemize{
\item \code{id}: parameter ID (ie. ETA1, ETA2, ..., ETAn)
\item \code{name}: parameter name
\item \code{value}: eta bar value
\item \code{se}: standard error
\item \code{n}: sample size
\item \code{pvalue}: p-value helping assess whether the sample average is "far" from 0
}
}
\description{
Summarize ETA bars of the individual parameters resulting of an estimation step.
}
\examples{
EXAMPLERUN \%>\% summarize_etabars()
}
