% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_population.R
\name{summarize_omega}
\alias{summarize_omega}
\title{Summarize OMEGA matrix}
\usage{
summarize_omega(run, estimation_number = NULL)
}
\value{
A data frame with one row per parameter and the following columns:
\itemize{
\item \code{eta1}: first parameter name
\item \code{eta2}: second parameter name
\item \code{estimate}: final estimate
\item \code{se}: standard error
\item \code{rse}: relative standard error
\item \code{ci_low}: lower endpoint of the 95\% confidence interval
\item \code{ci_up}: upper endpoint of the 95\% confidence interval
\item \code{cv}: coefficient of variation, considering the random parameter
is a variability term associated to a log-normally distributed parameter
}
}
\description{
Summarize variance-covariance matrix of the random parameters values resulting of an estimation step.
}
\examples{

EXAMPLERUN \%>\% summarize_omega()
}
