% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_parameters_correlations.R
\name{summarize_parameters_correlations}
\alias{summarize_parameters_correlations}
\title{Parameters correlations}
\usage{
summarize_parameters_correlations(run, parameters = NULL,
  baseline_only = TRUE, correlation_method = NULL, auto_order = TRUE)
}
\arguments{
\item{run}{\code{pmxploit} NONMEM run object.}

\item{parameters}{character vector of parameters names. Default is \code{NULL},
returning all individual parameters (random and post-hoc).}

\item{baseline_only}{logical. Consider only the baseline (= first) values of the subjects. Default is
\code{TRUE}.}

\item{correlation_method}{a character string indicating which correlation
coefficient (or covariance) is to be computed. One of \code{"pearson"}
(default), \code{"kendall"}, or \code{"spearman"}: can be abbreviated. If
\code{type = "scatterplot"}, \code{"pearson"} method will be used.}

\item{auto_order}{logical. If \code{type = "heatmap"}, auto-orders the
matrix with respect to the distances between values. Default is
\code{TRUE}.}
}
\value{
A matrix.
}
\description{
Returns the selected parameters correlations in the form of either a matrix, a
data frame or a plot.
}
\examples{
EXAMPLERUN \%>\% summarize_parameters_correlations()
EXAMPLERUN \%>\% summarize_parameters_correlations(auto_order = FALSE)
}
