% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_parameters_distributions.R
\name{summarize_parameters_distributions}
\alias{summarize_parameters_distributions}
\title{Individual parameters distributions summary}
\usage{
summarize_parameters_distributions(
  run,
  parameters = NULL,
  quantiles = c(0.05, 0.25, 0.75, 0.95),
  baseline_only = TRUE
)
}
\arguments{
\item{run}{\code{pmxploit} NONMEM run object.}

\item{parameters}{character vector of parameters names. Default is \code{NULL},
returning all individual parameters (random and post-hoc).}

\item{quantiles}{numeric vector of quantiles. Default are 5th, 25th, 75th and 95th percentiles.}

\item{baseline_only}{logical. Consider only the baseline (= first) values of the subjects. Default is
\code{TRUE}.}
}
\value{
A data frame.
}
\description{
Summarizes the selected individual parameters: number of values, mean, median,
quantiles and range.
}
\examples{
EXAMPLERUN \%>\% summarize_parameters_distributions()

EXAMPLERUN \%>\% summarize_parameters_distributions(quantiles = seq(0.05, 0.95, 0.05))
EXAMPLERUN \%>\% summarize_parameters_distributions(quantiles = NULL)

EXAMPLERUN \%>\% group_by(STUD) \%>\% summarize_parameters_distributions()
EXAMPLERUN \%>\% group_by(STUD) \%>\% summarize_parameters_distributions(quantiles = NULL)
}
