% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_parameters_vs_continuous_covariates.R
\name{summarize_parameters_vs_continuous_covariates}
\alias{summarize_parameters_vs_continuous_covariates}
\title{Parameters vs continuous covariates correlations}
\usage{
summarize_parameters_vs_continuous_covariates(
  run,
  parameters = NULL,
  covariates = NULL,
  baseline_only = TRUE,
  correlation_method = NULL,
  auto_order = TRUE
)
}
\arguments{
\item{run}{\code{pmxploit} NONMEM run object.}

\item{parameters}{character vector of parameters names. Default is \code{NULL},
returning all individual parameters (random and post-hoc).}

\item{covariates}{character vector of continous covariates names. Default is
\code{NULL}, returning all continuous covariates.}

\item{baseline_only}{logical. Consider only the baseline (= first) values of the subjects. Default is
\code{TRUE}.}

\item{correlation_method}{a character string indicating which correlation
coefficient (or covariance) is to be computed. One of \code{"pearson"}
(default), \code{"kendall"}, or \code{"spearman"}: can be abbreviated. If
\code{type = "scatterplot"}, \code{"pearson"} method will be used.}

\item{auto_order}{logical. If \code{type = "heatmap"}, auto-orders the
matrix with respect to the distances between values. Default is
\code{TRUE}.}
}
\value{
A matrix.
}
\description{
Returns a matrix of the selected parameters and continuous covariates correlations.
}
\examples{
EXAMPLERUN \%>\% summarize_parameters_vs_continuous_covariates()
}
