% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_population.R
\name{summarize_shrinkage}
\alias{summarize_shrinkage}
\title{Summarize shrinkage}
\usage{
summarize_shrinkage(run, estimation_number = NULL, type = NULL)
}
\arguments{
\item{run}{\code{pmxploit} NONMEM run object.}

\item{estimation_number}{integer. Number of the estimation step.
Default is NULL, returning the last estimation step.}

\item{type}{character vector. Shrinkage type (ETA, EBV or EPS).}
}
\value{
A data frame with one row per shrinkage value and the following columns:
\itemize{
\item \code{type}: shrinkage type (ETA, EBV or EPS)
\item \code{parameter}: random parameter name
\item \code{shrinkage}: shrinkage value
}
}
\description{
Summarize shrinkage (ETA, EBV and EPS) of the individual parameters resulting of an estimation step.
}
\examples{

EXAMPLERUN \%>\% summarize_shrinkage()
EXAMPLERUN \%>\% summarize_shrinkage(type = "ETA")
}
