% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_population.R
\name{summarize_sigma}
\alias{summarize_sigma}
\title{Summarize SIGMA matrix}
\usage{
summarize_sigma(run, estimation_number = NULL)
}
\arguments{
\item{run}{\code{pmxploit} NONMEM run object.}

\item{estimation_number}{integer. Number of the estimation step.
Default is NULL, returning the last estimation step.}
}
\value{
A data frame with one row per parameter and the following columns:
\itemize{
\item \code{epsilon1}: first parameter name
\item \code{epsilon2}: second parameter name
\item \code{estimate}: final estimate
\item \code{se}: standard error
\item \code{rse}: relative standard error
\item \code{ci_low}: lower endpoint of the 95\% confidence interval
\item \code{ci_up}: upper endpoint of the 95\% confidence interval
}
}
\description{
Summarize variance-covariance matrix of the residual error parameters values resulting of an estimation step.
}
\examples{

EXAMPLERUN \%>\% summarize_sigma()
}
