% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_population.R
\name{summarize_thetas}
\alias{summarize_thetas}
\title{Summarize THETA parameters}
\usage{
summarize_thetas(run, estimation_number = NULL)
}
\arguments{
\item{run}{\code{pmxploit} NONMEM run object.}

\item{estimation_number}{integer. Number of the estimation step.
Default is NULL, returning the last estimation step.}
}
\value{
A data frame with one row per parameter and the following columns:
\itemize{
\item \code{id}: parameter ID (ie. THETA1, THETA2, ..., THETAn)
\item \code{name}: parameter name
\item \code{estimate}: final estimate
\item \code{se}: standard error
\item \code{rse}: relative standard error
\item \code{ci_low}: lower endpoint of the 95\% confidence interval
\item \code{ci_up}: upper endpoint of the 95\% confidence interval
}
}
\description{
Summarize typical population parameters values resulting of an estimation step.
}
\examples{

EXAMPLERUN \%>\% summarize_thetas()
}
