% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{summary.nonmem_control_stream}
\alias{summary.nonmem_control_stream}
\title{Print the summary of a NONMEM control stream object}
\usage{
\method{summary}{nonmem_control_stream}(x, ...)
}
\arguments{
\item{x}{A \code{nonmem_control_stream} object generated.}
}
\description{
Print the summary of a NONMEM control stream object
}
\examples{
summary(EXAMPLERUN$control_stream)

}
