% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{switch_covariates_type}
\alias{switch_covariates_type}
\title{Switch covariate type (categorical <-> continuous)}
\usage{
switch_covariates_type(run, covariates = NULL)
}
\arguments{
\item{run}{\code{pmxploit} NONMEM run object.}

\item{covariates}{character vector of covariates names.}
}
\value{
A NONMEM run object.
}
\description{
Convert categorical covariates to continuous covariates and vice versa.
}
\examples{
switched_EXAMPLERUN <- switch_covariates_type(run = EXAMPLERUN, covariates = c("WT", "STUD"))

EXAMPLERUN \%>\% plot_continuous_covariates_distributions()
switched_EXAMPLERUN \%>\% plot_continuous_covariates_distributions()

EXAMPLERUN \%>\% plot_categorical_covariates_distributions()
}
