% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_pmx}
\alias{theme_pmx}
\title{\code{pmxploit} ggplot2 theme}
\usage{
theme_pmx()
}
\value{
A ggplot2 theme.
}
\description{
A green & white theme
}
\examples{
EXAMPLERUN \%>\%
     plot_residuals(compartment = 2, idv = c("TIME", "PRED"), residuals = "CWRES")+
     theme_pmx()
}
