% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_methods.R
\name{ungroup.nonmem_run}
\alias{ungroup.nonmem_run}
\title{Ungroup a NONMEM run object data}
\usage{
\method{ungroup}{nonmem_run}(x)
}
\arguments{
\item{x}{An \code{nonmem_run} object generated.}
}
\description{
Ungroup a NONMEM run object data
}
\examples{
ungroup(EXAMPLERUN)

}
