% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_nm_run.R
\name{load_nm_run_directory}
\alias{load_nm_run_directory}
\title{Load a NONMEM run data from a directory}
\usage{
load_nm_run_directory(path, dataset_separator = NA,
  dataset_na_string = ".", read_initial_values = TRUE,
  load_tables = TRUE, control_stream = NULL, update_progress = NULL,
  verbose = FALSE)
}
\arguments{
\item{control_stream}{\code{nonmem_control_stream} object. Control stream data, obtained from a prior
call to \code{parse_nm_control_stream()} function.}
}
\description{
Load a NONMEM run data from a directory
}
